LNP-1607: ♻️   move warmPrimaryNavigation() to the base class - there is nothing specific to the old or new implementations.
