+++ docroot/modules/custom/prisoner_hub_cache_warmer/src/Plugin/warmer/LegacyPrisonerHubWarmer.php
@@ -57,42 +57,6 @@ class LegacyPrisonerHubWarmer extends PrisonerHubWarmerBase {
   /**
    * {@inheritdoc}
    */
-  protected function warmPrimaryNavigation(string $prison) {
-    return $this->warmJsonApiRequestAsync("$this->cacheWarmerEndpoint/en/jsonapi/prison/$prison/primary_navigation?fields%5Bmenu_link_content--menu_link_content%5D=id%2Ctitle%2Curl")
-      ->then(function (ResponseInterface $response) use ($prison) {
-        if (!$json_response = json_decode($response->getBody())) {
-          return;
-        }
-        $tids = [];
-        foreach ($json_response->data as $menu_item) {
-          if (!isset($menu_item->attributes->url)) {
-            continue;
-          }
-          $matches = [];
-          if (preg_match("/tags\/(\d+)/", $menu_item->attributes->url, $matches)) {
-            $tids[] = $matches[1];
-          }
-        }
-        $terms = $this->termStorage->loadMultiple($tids);
-        foreach ($terms as $term) {
-          if ($term->bundle() != 'moj_categories') {
-            continue;
-          }
-          $this->warmCategoryPage($prison, $term->uuid());
-        }
-        foreach ($tids as $tid) {
-          $this->queueAsynchronousRouterRequest($prison, "translate-path?path=tags/$tid");
-        }
-
-      }, function (\Exception $e) {
-        Error::logException($this->logger, $e);
-      }
-      );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   protected function warmSeriesPage(string $prison, string $uuid) {
     $this->queueAsynchronousJsonApiRequest($prison, "node?filter%5Bfield_moj_series.id%5D=$uuid&include=field_moj_thumbnail_image%2Cfield_moj_series.field_moj_thumbnail_image&sort=series_sort_value%2Ccreated&fields%5Bnode--page%5D=drupal_internal__nid%2Ctitle%2Cfield_summary%2Cfield_moj_thumbnail_image%2Cfield_moj_series%2Cpath%2Cpublished_at&fields%5Bnode--moj_video_item%5D=drupal_internal__nid%2Ctitle%2Cfield_summary%2Cfield_moj_thumbnail_image%2Cfield_moj_series%2Cpath%2Cpublished_at&fields%5Bnode--moj_radio_item%5D=drupal_internal__nid%2Ctitle%2Cfield_summary%2Cfield_moj_thumbnail_image%2Cfield_moj_series%2Cpath%2Cpublished_at&fields%5Bnode--moj_pdf_item%5D=drupal_internal__nid%2Ctitle%2Cfield_summary%2Cfield_moj_thumbnail_image%2Cfield_moj_series%2Cpath%2Cpublished_at&fields%5Bfile--file%5D=image_style_uri&fields%5Btaxonomy_term--series%5D=name%2Cdescription%2Cdrupal_internal__tid%2Cfield_moj_thumbnail_image%2Cpath%2Cfield_exclude_feedback%2Cbreadcrumbs&page[offset]=0&page[limit]=40");
   }
